+++ tools/src/tester/ChapterEvaluator.ts
@@ -30,10 +30,11 @@ export default class ChapterEvaluator {
     this.schema_validator = SharedResources.get_instance().schema_validator
   }
 
-  async evaluate (skipped: boolean): Promise<ChapterEvaluation> {
-    if (skipped) return { title: this.chapter.synopsis, overall: { result: Result.SKIPPED } }
-    const operation = this.spec_parser.locate_operation(this.chapter)
+  async evaluate (skip: boolean): Promise<ChapterEvaluation> {
+    if (skip) return { title: this.chapter.synopsis, overall: { result: Result.SKIPPED } }
     const response = await this.chapter_reader.read(this.chapter)
+    const operation = this.spec_parser.locate_operation(this.chapter)
+    if (operation == null) return { title: this.chapter.synopsis, overall: { result: Result.FAILED, message: `Operation "${this.chapter.method.toUpperCase()} ${this.chapter.path}" not found in the spec.` } }
     const params = this.#evaluate_parameters(operation)
     const request_body = this.#evaluate_request_body(operation)
     const status = this.#evaluate_status(response)
@@ -41,7 +42,7 @@ export default class ChapterEvaluator {
     return {
       title: this.chapter.synopsis,
       overall: { result: overall_result(Object.values(params).concat([request_body, status, payload])) },
-      request: { parameters: params, requestBody: request_body },
+      request: { parameters: params, request_body },
       response: { status, payload }
     }
   }